MONDRIAN-PACINO: Makes the BasicQueryTest.testconcurrentStatementRun less agressive. Instead of doing an uber-crossjoin, it now does a TopCount instead.
Also fixes an issue with the AggregationManager sqlExecutor. It was using a ThreadPoolExecutor backed by a limited queue of 10 slots, and that limit was very easy to attain by sending a few concurrent statements to a single server. I have replaced the executor by an expanding cached thread pool which uses 10 threads at maximum, but uses an expanding queue to back it.

[git-p4: depot-paths = "//open/mondrian-release/pacino/": change = 14788]