MONDRIAN-PACINO: Fixes an issue with composite batches. Certain code paths would not always return future objects to the FastBatchingCellReader.loadAggregations method so iteration limit would kick in before the segments has time to load.
Also fixes an issue with InlineTableTest. It had a special case handling for Oracle and Access where it would assume that the Access and Oracle dialects were returning timestamps instead of dates for columns that are marked as the Date type. This was a bug that was fixed in Mondrian 3.3.0 but was not ported yet to the Access dialect as well.

[git-p4: depot-paths = "//open/mondrian-release/pacino/": change = 14822]